requirements.txt
fastapi
uvicorn
python-dotenv
pinecone-client
langchain-openai
PyPDF2